requirements.txt
weaviate-client~=4.6.3
streamlit~=1.35.0
python-dotenv~=1.0.0
pytest~=7.4.0
isort~=5.12.0
pytest-cov~=4.1.0
ipykernel~=6.25.0
pytest-docker~=2.0.0
wasabi~=1.1.2
typer~=0.9.0
